requirements.txt
requests>=2.31.0
icalendar>=5.0.0
python-dateutil>=2.8.2
PyYAML>=6.0.1
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.2
pycountry>=22.3.5
geopy>=2.4.1
unidecode>=1.3.7
python-dotenv>=1.0.0
